правила лаб1 и перенесла задачи codewars
