+++ golang/lab1.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"fmt"
 	"math"
 )
 
@@ -10,18 +9,12 @@ func function(a, x float64) float64 {
 	return y
 }
 
-func zadacha1(a float64) float64 {
-	fmt.Println("Задание А")
-	for x := 1.2; x <= 3.7; x += 0.5 {
-		fmt.Println("x = ", x, " ", "y = ", function(a, x))
-	}
-	return (1)
+func zadacha1(a, x float64) float64 {
+	var f float64 = function(a, x)
+	return f
 }
-func zadacha2(a float64) float64 {
-	var znach = [5]float64{1.28, 1.36, 2.47, 3.68, 4.56}
-	fmt.Println("Задание B")
-	for _, x1 := range znach {
-		fmt.Println("x = ", x1, " ", "y = ", function(a, x1))
-	}
-	return (1)
+
+func zadacha2(a, x1 float64) float64 {
+	var f1 float64 = function(a, x1)
+	return f1
 }
